Update footer.html
Continuing link.yaml fix,

Due to issue at footer.html when clicking on the link at the bottom of each page at the "Find an issue with this page? Fix it on GitHub" Link forwarded to a wierd URL.

changed the <a> tag to fix the link to submit a new issue in github.