Update links.yaml
Continuing link.yaml fix,

Due to issue at footer.html when clicking on the link at the bottom of each page at the "Find an issue with this page? Fix it on GitHub" Link forwarded to a wierd URL.

added github_submit_fixes to be used in footer.html to direct to a correct url when submitting a fix.