Help test Rust version of GUN???
Pinged you on Twitter, thought I'd make it easy for accept/reject: In the next versions of GUN the community approved of running an async & delayed A/B test at the bottom of gun.js, it'll websocket connect to a Rust and/or JS "relay" peer to read 1 test value and measure latency. If you're willing to help, I'd actually like to learn what Firebase calls you make and make mock ones to GUN and then compare 99th percentile latencies, so I can keep on improving until as good as Firebase.