Fix shell command
The description of this command in deploy-multiple-sites-to-firebase-hosting says it 'deploys a single site based on the target name', but 'firebase serve' will only allow the site to be served locally.

I believe the intended command was 'firebase deploy', which will deploy only a single site based on the target name, as described by the description above.