Fixes #1487 - Autofocus search modal input
This commit addresses issue #1487 by adding a proper autofocus on algolia-search modal. This focus on the input field when user types something for the first time.
Pull requests ##1418 add autofocus on search input, but that doesn't seem to be working. Also, this solution produced a warning "Avoid using autofocus".
Pull request #1442 also resolves this issue but uses setTimeout to autofocus.