Update classes-basic.md
Hey, the dart course is a gem! 
Currently, there is a missing keyword on line 49. It causes the following error

```
Variables must be declared using the keywords 'const', 'final', 'var', or a type name.
Try adding the name of the type of the variable or the keyword 'var'.dart(missing_const_final_var_or_type)
```

This should fix it